Order api

See merge request b20100/fastfood/app-order-parent!14

+++ app-order/src/main/java/uz/pdp/apporder/payload/OrderProductsDTO.java
@@ -16,6 +16,6 @@ public class OrderProductsDTO {
     private Integer productId;
 
     @NotNull
-    private Float quantity;
+    private Short quantity;
 
 }
